reduce used heap space and started with normalisation

+++ src/CNNtest.java
@@ -6,12 +6,12 @@ public class CNNtest extends Structure {
     public CNNtest(){
     }
 
-    public static double CNNtest(String[][] architecture, network convnet, double[][][]test_x, int[]test_y){
+    public static float CNNtest(String[][] architecture, network convnet, float[][][]test_x, int[]test_y){
 
         CNNff.CNNff(architecture,convnet,test_x);
         int[] output = new int[convnet.o[0].length];
         for (int i = 0; i < convnet.o[0].length; i++) {
-            double max = 0;
+            float max = 0;
             int j_max = 0;
             for (int j = 0; j < convnet.o.length; j++) {
                 if (convnet.o[j][i]>max){
@@ -28,7 +28,7 @@ public class CNNtest extends Structure {
                 correct++;
         }
 
-        return (double) correct/test_y.length;
+        return (float) correct/test_y.length;
     }
 
 }
